backend/python-services/heightmap-service/requirements.txt
pika>=1.3.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0
boto3>=1.34.0
prometheus-client>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
scipy>=1.11.0
python-json-logger>=2.0.0
GDAL==3.10.3
requests>=2.31.0